Turned off server side rendering for dev mode to allow hot reloading of CSS. Fix tests and add mocks for jest to ignore styles and files.
