+++ .babelrc
@@ -6,5 +6,17 @@
   ],
   "plugins": [
     "transform-flow-strip-types"
-  ]
+  ],
+  "env": {
+    "node": {
+      "plugins": [
+        [
+          "babel-plugin-transform-require-ignore",
+          {
+            "extensions": [".css", ".scss"]
+          }
+        ]
+      ]
+    }
+  }
 }
