+++ package.json
@@ -25,6 +25,7 @@
     "babel-eslint": "^7.1.1",
     "babel-loader": "^7.0.0",
     "babel-plugin-transform-flow-strip-types": "^6.22.0",
+    "babel-plugin-transform-require-ignore": "^0.1.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-2": "^6.24.1",
@@ -63,7 +64,7 @@
     "flow:update": "flow-typed install",
     "build": "npm-run-all --parallel build:server build:client",
     "prebuild:server": "rimraf build",
-    "build:server": "babel src -d build --source-maps --ignore test.js",
+    "build:server": "BABEL_ENV=node babel src -d build --copy-files --source-maps --ignore test.js",
     "prebuild:client": "rimraf public/build",
     "build:client": "cross-env NODE_ENV=production webpack --progress --env=dist",
     "prepush": "npm run validate"
@@ -81,6 +82,10 @@
   "homepage": "https://github.com/ericmasiello/express-starter#readme",
   "description": "",
   "jest": {
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js"
+    },
     "collectCoverageFrom": [
       "src/server/**/*.js"
     ],
