+++ src/client/Detail.css
@@ -1,0 +1,3 @@
+.detail {
+  color: darkblue;
+}
