+++ src/client/Home.js
@@ -1,8 +1,8 @@
 /* @flow */
 import React, { Component } from 'react';
+import { Link } from 'react-router';
 import type { Namespace$AppState } from '../types';
-
-if(process.env.WEBPACK) require('./Home.css');
+import './Home.css';
 
 export default class Home extends Component {
   constructor() {
@@ -19,6 +19,9 @@ export default class Home extends Component {
       <div>
         <h2>Home page! {this.state.toggle ? 'Yup!' : 'Nope'}</h2>
         <button onClick={() => this.setState({ toggle: !this.state.toggle })}>Toggle</button>
+        <div>
+          <Link to="detail/1234">Details</Link>
+        </div>
       </div>
     );
   }
