+++ src/server/util/routeMatch.js
@@ -2,8 +2,8 @@
 import React from 'react';
 import { renderToString } from 'react-dom/server';
 import { RouterContext, match } from 'react-router';
+// import { match } from 'react-router';
 import logger from '../logger';
-import routes from '../../client/routes';
 import isprod from './isprod';
 import configFactory from '../../../webpack.config';
 import type { Namespace$WebpackConfig } from '../../types';
@@ -52,8 +52,19 @@ export function routeMatchCallback(response: express$Response): Function {
 }
 
 export default function routeMatch(request: express$Request, response: express$Response): void {
-  return match({
-    routes,
-    location: request.url,
-  }, routeMatchCallback(response));
+  if (isprod) {
+    const routes = require('../../client/routes').default; // eslint-disable-line global-require
+    match({
+      routes,
+      location: request.url,
+    }, routeMatchCallback(response));
+    return;
+  }
+
+  const buildPath = getBuildPath(isprod, distConfig);
+
+  response.render('index', {
+    html: '',
+    buildPath,
+  });
 }
